Module stdsimd::arch::arm

🔬 This is a nightly-only experimental API. (stdsimd)

This is supported on ARM only.

Platform-specific intrinsics for the arm platform.

See the module documentation for more details.

Structs

int16x2_t	[
Experimental
] [
ARM
]
ARM-specific 32-bit wide vector of two packed i16.

int8x4_t	[
Experimental
] [
ARM
]
ARM-specific 32-bit wide vector of four packed i8.

uint16x2_t	[
Experimental
] [
ARM
]
ARM-specific 32-bit wide vector of two packed u16.

uint8x4_t	[
Experimental
] [
ARM
]
ARM-specific 32-bit wide vector of four packed u8.

Functions

__breakpoint⚠	[
Experimental
] [
ARM
]
Inserts a breakpoint instruction.

_clz_u8⚠	[
Experimental
] [
ARM and v7
]
Count Leading Zeros.

_clz_u16⚠	[
Experimental
] [
ARM and v7
]
Count Leading Zeros.

_clz_u32⚠	[
Experimental
] [
ARM and v7
]
Count Leading Zeros.

_rbit_u32⚠	[
Experimental
] [
ARM and v7
]
Reverse the bit order.

_rev_u16⚠	[
Experimental
] [
ARM
]
Reverse the order of the bytes.

_rev_u32⚠	[
Experimental
] [
ARM
]
Reverse the order of the bytes.

qadd⚠	[
Experimental
] [
ARM
]
Signed saturating addition

qadd8⚠	[
Experimental
] [
ARM
]
Saturating four 8-bit integer additions

qadd16⚠	[
Experimental
] [
ARM
]
Saturating two 16-bit integer additions

qasx⚠	[
Experimental
] [
ARM
]
Returns the 16-bit signed saturated equivalent of

qsax⚠	[
Experimental
] [
ARM
]
Returns the 16-bit signed saturated equivalent of

qsub⚠	[
Experimental
] [
ARM
]
Signed saturating subtraction

qsub8⚠	[
Experimental
] [
ARM
]
Saturating two 8-bit integer subtraction

qsub16⚠	[
Experimental
] [
ARM
]
Saturating two 16-bit integer subtraction

sadd8⚠	[
Experimental
] [
ARM
]
Returns the 8-bit signed saturated equivalent of

sadd16⚠	[
Experimental
] [
ARM
]
Returns the 16-bit signed saturated equivalent of

sasx⚠	[
Experimental
] [
ARM
]
Returns the 16-bit signed equivalent of

sel⚠	[
Experimental
] [
ARM
]
Select bytes from each operand according to APSR GE flags

shadd8⚠	[
Experimental
] [
ARM
]
Signed halving parallel byte-wise addition.

shadd16⚠	[
Experimental
] [
ARM
]
Signed halving parallel halfword-wise addition.

shsub8⚠	[
Experimental
] [
ARM
]
Signed halving parallel byte-wise subtraction.

shsub16⚠	[
Experimental
] [
ARM
]
Signed halving parallel halfword-wise subtraction.

smlad⚠	[
Experimental
] [
ARM
]
Dual 16-bit Signed Multiply with Addition of products and 32-bit accumulation.

smlsd⚠	[
Experimental
] [
ARM
]
Dual 16-bit Signed Multiply with Subtraction of products and 32-bit accumulation and overflow detection.

smuad⚠	[
Experimental
] [
ARM
]
Signed Dual Multiply Add.

smuadx⚠	[
Experimental
] [
ARM
]
Signed Dual Multiply Add Reversed.

smusd⚠	[
Experimental
] [
ARM
]
Signed Dual Multiply Subtract.

smusdx⚠	[
Experimental
] [
ARM
]
Signed Dual Multiply Subtract Reversed.

udf⚠	[
Experimental
] [
ARM
]
Generates the trap instruction UDF

usad8⚠	[
Experimental
] [
ARM
]
Sum of 8-bit absolute differences.

usad8a⚠	[
Experimental
] [
ARM
]
Sum of 8-bit absolute differences and constant.